Revert "[BKR-623] Reorganize Beaker test runner classes for introduction of minitest runner"
