+++ lib/beaker/cli.rb
@@ -154,13 +154,7 @@ module Beaker
         @logger.notify("No tests to run for suite '#{suite_name.to_s}'")
         return
       end
-
-      unless runner_class = Beaker::TestSuite.runner(@options[:runner])
-        @logger.error "Test runner #{@options[:runner]} is unknown."
-        exit 1
-      end
-
-      runner_class.new(
+      Beaker::TestSuite.new(
         suite_name, @hosts, @options, @timestamp, failure_strategy
       ).run_and_raise_on_failure
     end
